complete

+++ php/post/comments.php
@@ -1,11 +1,11 @@
 <?php
 session_start();
-require_once 'conn.php';
+require_once '../conn.php';
 if(isset($_POST['comment']))
 {
-	if($_POST['comment']== 1)
+	if($_POST['comment']== 'create comment')
 	{
-		$comment_id = mysqli_real_escape_string($link, $_SESSION['user_id']);
+		$comment_id = mysqli_real_escape_string($link, $_SESSION['sess_id']);
 		$post_id = mysqli_real_escape_string($link, $_POST['post_id']);
 		$comment = mysqli_real_escape_string($link, $_POST['content']);
 		date_default_timezone_set("Asia/Calcutta");
@@ -13,7 +13,7 @@ if(isset($_POST['comment']))
 		$query = "INSERT INTO `comments` (`comment_id`, `post_id`,`comment`,`date_created`) VALUES ('$comment_id', '$post_id', '$comment','$date_now')";
 		if($result = mysqli_query($link, $query))
 		{  
-		    $data['name'] = $_SESSION['name']; 
+		    $data['name'] = $_SESSION['user_name']; 
 		    $data['content'] = $comment;
 			$data['date_created'] = $date_now;
 			$data['status'] = 201;
@@ -26,7 +26,7 @@ if(isset($_POST['comment']))
 			echo json_encode($data);
 		}
 	}
-	else if($_POST['comment']== 2)
+	else if($_POST['comment']== 'fetch comment')
 	{
 	    $post_id = mysqli_real_escape_string($link, $_POST['post_id']);	
 		$query = "SELECT * FROM `comments` WHERE `post_id` = '$post_id'";
@@ -36,9 +36,9 @@ if(isset($_POST['comment']))
 		    $data = array(array());
 		    while($row = $result->fetch_assoc())
 			{
-				$data[$i]['user_id'] = $row['comment_id'];
-				$data[$i]['comments'] = $row['comment'];
-				$data[$i]['date_create'] = $row['date_created'];
+				$data[$i]['id'] = $row['comment_id'];
+				$data[$i]['comments'] = $row['content'];
+				$data[$i]['date_created'] = $row['date_created'];
 				$i++;
 		    }
 			echo json_encode($data);
@@ -51,11 +51,11 @@ if(isset($_POST['comment']))
 			echo json_encode($data);
 		}
 	}
-    else 
+    else if($_POST['comment']== 'delete comment') 
     {
-		$comment_id = mysqli_real_escape_string($link, $_SESSION['user_id']);
+		$comment_id = mysqli_real_escape_string($link, $_SESSION['sess_id']);
 		$post_id = mysqli_real_escape_string($link, $_POST['post_id']);
-		$date = mysqli_real_escape_string($link, $_POST['time']);
+		$date = mysqli_real_escape_string($link, $_POST['date_created']);
 		$query = "DELETE FROM `comments` WHERE `comment_id` = '$comment_id' AND `post_id` = '$post_id' AND`date_created` = '$date'";
 		if($result = mysqli_query($link, $query))
 		{  
@@ -69,7 +69,34 @@ if(isset($_POST['comment']))
 			echo json_encode($data);
 		}
 		
+	}
+	 else if($_POST['comment']== 'update comment') 
+    {
+		$comment_id = mysqli_real_escape_string($link, $_SESSION['sess_id']);
+		$post_id = mysqli_real_escape_string($link, $_POST['post_id']);
+		$comment = mysqli_real_escape_string($link, $_POST['content']);
+		$date = mysqli_real_escape_string($link, $_POST['date_created']);
+		date_default_timezone_set("Asia/Calcutta");
+		$date_now = date("r");
+	    $query = "UPDATE `comments` SET `date_created` = '$date_now' , `comment` = '$comment' WHERE `comment_id` = '$comment_id' AND `post_id` = '$post_id' AND`date_created` = '$date'";	
+	    if($result = mysqli_query($link, $query))
+		{  
+		    $data['content'] = $comment;
+			$data['date_created'] = $date_now;
+			$data['status'] = 201;
+			echo json_encode($data);
+		}  
+		else  
+		{  
+			$data['status'] = 601;
+			$data['error'] = $link -> error;
+			echo json_encode($data);
+		}
 	}	
+    else
+    {
+	    header('Location: http://index.php');
+    } 	
 }
 else
 {
